Update policy.md
The suggested language is mandated by federal law, both in terms of the U.S. code and appropriations restrictions. There is no discretion about including it, or any associated spending to implement or enforce the policy would be illegal. Inclusion of the addendum is required by Section 104 of the Whistleblower Protection Enhancement Act, Section 115 of the Whistleblower Protection Enhancement Act, and Section 743, Division E of the Consolidated Appropriations Act for FY 2015.